refactor(ui): extract conversation input field into separate component
